Changed whiteAdvantage back to a double. Added genetic alg library and old games as training.

+++ TakLib/AI/BoardAnalysisData.cs
@@ -76,7 +76,7 @@ namespace TakLib
 
         public BoardAnalysisWeights weights;
 
-        public int whiteAdvantage { get; set; }
+        public double whiteAdvantage { get; set; }
 
         public BoardAnalysisData() { }
